requirements.txt
adblockparser==0.7
arrow==0.12.1
asn1crypto==0.24.0
astroid==2.1.0
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.3
bleach==3.1.0
boto==2.49.0
boto3==1.9.71
botocore==1.12.71
bz2file==0.98
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
click-plugins==1.0.4
cligj==0.5.0
cloudpickle==0.8.0
colorama==0.4.1
dnspython==1.16.0
docopt==0.6.2
docutils==0.14
dragnet==2.0.4
editdistance==0.5.2
Fiona==1.8.4
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
ftfy==4.4.3
gensim==3.6.0
gevent==1.3.7
graphviz==0.8.4
greenlet==0.4.15
hdfs==2.5.1
HeapDict==1.0.0
html5lib==1.0.1
httplib2==0.12.0
hyperlink==17.3.1
idna==2.8
incremental==16.10.1
iotop==0.6
isort==4.3.4
itsdangerous==1.1.0
jellyfish==0.7.1
Jinja2==2.10
jmespath==0.9.3
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
lazy-object-proxy==1.3.1
locket==0.2.0
lxml==4.3.1
lz4==2.1.6
MarkupSafe==1.1.0
matplotlib==3.0.2
maxminddb==1.4.1
mccabe==0.6.1
msgpack==0.6.1
munch==2.3.2
mxnet==1.1.0.post0
netifaces==0.10.4
networkx==2.2
nltk==3.4
numpy==1.15.4
oauthlib==2.1.0
PAM==0.4.2
pandas==0.23.4
partd==0.3.9
passlib==1.7.1
patsy==0.5.1
Pillow==5.4.1
prettytable==0.7.2
progressbar2==3.39.3
protobuf==3.6.1
psutil==5.6.0
psycopg2==2.7.6.1
pyarrow==0.11.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pycurl==7.43.0.1
pygobject==3.26.1
pylint==2.2.2
pyOpenSSL==17.5.0
pyparsing==2.3.0
Pyro4==4.74
pyserial==3.4
PySocks==1.6.8
pyspeedtest==1.2.7
python-apt==1.6.3
python-dateutil==2.8.0
python-debian==0.1.32
python-louvain==0.13
python-simhash==0.1
python-utils==2.3.0
pythonwhois==2.4.3
pytz==2018.7
pyxdg==0.25
PyYAML==3.13
re2==0.2.22
redis==3.0.1
reppy==0.4.12
requests==2.21.0
requests-file==1.4.3
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
s3transfer==0.1.13
scapy==2.4.0
scikit-learn==0.20.2
scipy==1.2.0
scour==0.36
seaborn==0.9.0
SecretStorage==2.3.1
serpent==1.27
service-identity==16.0.0
Shapely==1.6.4.post2
simhash==1.9.0
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.7.1
sortedcontainers==2.1.0
speedtest-cli==2.1.1
SQLAlchemy==1.2.15
ssh-import-id==5.7
statsmodels==0.9.0
strsim==0.0.3
systemd-python==234
tblib==1.3.2
termcolor==1.1.0
tld==0.9.1
tldextract==2.2.0
toolz==0.9.0
tornado==5.1.1
turicreate==5.3.1
tweepy==3.7.0
Twisted==17.9.0
twitter==1.18.0
typed-ast==1.1.1
tzdata==1.2016.8
ufw==0.35
ujson==1.35
unattended-upgrades==0.1
urllib3==1.24.1
virtualenv==16.2.0
wcwidth==0.1.7
webencodings==0.5.1
websocket==0.2.1
Werkzeug==0.14.1
wrapt==1.10.11
zict==0.1.3
zope.interface==4.3.2
